Report lag calculated at requst time
Lags in Burrow are evaluated at the time of the offset commit. Stopped consumers may
still report 0 lag, as it was calculated when the last commit arrived. This can be confusing.

This change calculates the lag during the evaluation time by comparing broker
offset with the latest consumer offset received for the partition.